refactoring folders and files
